spelling and grammar check
In the section "straightforward 51% attack" there's a sentence I don't understand:
"30-50% price drop for PoS coin, that is approximately 20% of the total supply."
What does 20% of total supply have to do with the example? Do you mean that selling 20% of the total supply leads to a 30-50% drop in price?

In the section "short selling attack", what is "comprometation"? I've never heard that word and it's not in any dictionary I found.